Added user profile
created the user profile. profile.css is empty for now, but will
eventually be used with this to add some color to the design. The
design could use more work, but I’m done with it for now, as dealing
with ejs to dynamically generate reviews ate up too much time.