Fixed the one issue. turns out it was browser specific and only applied on local.
Also, The system now can catch ejs stuff perfectly. It's passed to a service coded directly into index.ejs, which you can add to any controller to use the contents passed from the server to the current page. the only potential issue is when the page didn't refresh, but that's easy to fix.