Fixed things. now you can define the title of a page! put it in the state call object as the property 'title'.
At some point in the future, we might want to generate additions to MyApp.js and index.js based on directory names so we don't need to keep adding the new files we create to each. it would be pretty simple if we just add the title later; all the information is there.