Whoo! All pages work correctly now.
PPVCache is integrated into the controllers. we're just waiting on Shahbaz to set up spark and the PPV. Need him to maintain id though for a given ppvcache item. I really don't want to have to implement a button when the value is right there. Also need to get Atif to redesign the UI.